PoC: Updated Placeholder struct in devrantbee.gio
This is more like a proof on concept.
'Generating' the []bees.Placeholder struct with the 'reflect' package.
Unfortunately I'm violating the beehive style standard with changing
some PlaceholderDescriptor names like 'ID' to 'Id'.
The PoC shows that we don't have to write all the Placeholder structs
in beehive per hand