Configuration API
* Move json config related tasks to its own package (cfg)
* Provides an API to get/set the configuration (/v1/config endpoint)
* Saves the config every time we change it, instead of waiting for a graceful
  shutdown. This prevents losing config changes for unclean daemon shutdowns.

One of the current limitations of this approach is that it has been designed
with a single API client in mind, so saving the config isn't safe in a
multi-client environment.
If several clients try to update the config at the same time, only one will
succeed and changes from other clients will be discarded.

I'm not sure how big of a problem that'd be for beehive in reality.