Bees can now create Events
Bees can now create Events based on the name and a map containing the
Placeholders with Names and Values.
To do this we use the Event&Placeholder-Descriptor information which given in
the Bees Factory Namespace.
This will lead to a much shorter declaration of occuring Events in the Bees
implementation.