Load and write config from/to standard paths
This changes Beehive's behavior.

When writting the config file for the first time
(i.e. if beehive was run without a config), the file will be written to
`~/.config/beehive/beehive.conf` (OS dependant) instead of the current
directory.

Loading the config from the current directory, if available, is still
preferred, so that doesn't change.

Fixes #283