openweathermap: Handle multiple locations in context
Extends https://github.com/muesli/beehive/pull/311 by adding support
to store multiple locations in the Bee context, at the expense of making
it a bit more complext to retrieve the info from a template.

Something like:

"Weather in {{(index .context.weather.current 3128760).Name}}: ..."

Where 3128760 is the OpenWeatherMap ID corresponding to one of the
locations stored in the context while retrieving the weather.